% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_readCounting.R
\name{appendIntronicCounts}
\alias{appendIntronicCounts}
\title{Internal function appendIntronicCounts()}
\usage{
appendIntronicCounts(bamFile, saveDir, libraryType, verbose = FALSE)
}
\arguments{
\item{bamFile}{Path to read alignments in a bam file.}

\item{saveDir}{Name for the Project Directory will be saved.
The directory should contain a file "curated_introns.txt", as output by function ['curateIntrons()'].}

\item{libraryType}{Read type. Either "SE" (single-end) or "PE" (paired-end)}

\item{verbose}{logical indicating if verbose information for ['featureCount()'] debugging will be generated.}
}
\value{
Intron read counts in a file names "intronic_counts.txt", saved in directory \code{saveDir}.
}
\description{
This function counts reads overlapping
independent introns intervals using ['Rsubread::featureCount()'].
These counts will be merged to intron intervals provided in file "curated_introns.txt".
}
