% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_readCounting.R
\name{bootstrapJunctionCounts}
\alias{bootstrapJunctionCounts}
\title{Internal function bootstrapJunctionCounts()}
\usage{
bootstrapJunctionCounts(
  bamFile,
  saveDir,
  libraryType,
  nboot = 100,
  verbose = FALSE
)
}
\arguments{
\item{bamFile}{Path to read alignments in a bam file.}

\item{saveDir}{Working project directory.}

\item{libraryType}{Type of the library of RNA-seq data. Either "SE" (single-end) or "PE" (paired-end).}

\item{nboot}{Number of bootstrap sample to draw (Default: 100).}

\item{verbose}{logical indicating if verbose information for ['featureCount()'] debugging will be generated.}
}
\value{
Create several files with bootstrap results
}
\description{
This function computes a bootstrap estimate of gene expression by using exon-exon junction counts.
It is not intended to be used alone, but as a part of the IRratio2 estimation procedure,
wrapped in function ['computeIRratio2()'].
}
\seealso{
['featureCount()'], ['computeIRratio2()']
}
