% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_readCounting.R
\name{bootstrapMean}
\alias{bootstrapMean}
\title{Internal function bootstrapMean()}
\usage{
bootstrapMean(SpliceCounts, ExonToIntronCounts, nboot = 100, sboot = 50)
}
\arguments{
\item{SpliceCounts}{Vector with exon to exon counts.}

\item{ExonToIntronCounts}{Vector with exon to intron counts. Should have same size as \code{SpliceCounts}.}

\item{nboot}{Number of bootstrap sample to draw (Default: 100).}

\item{sboot}{Size of bootstrap samples (Default 50).}
}
\value{

}
\description{
This function does something useful
}
