% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_readCounting.R
\name{bootstrapSample}
\alias{bootstrapSample}
\title{Internal function bootstrapSample()}
\usage{
bootstrapSample(counts, sboot)
}
\arguments{
\item{counts}{Vector of counts to sample from.}

\item{sboot}{Size of bootstrap samples.}
}
\value{
A 2-vector with the mean and variance of a sample of size \code{sboot} drawn with replacement in \code{counts}.
}
\description{
This function simply creates a bootstrap sample from a vector of counts,
and computes the sample mean and variance.
}
