% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_SIRratio_wrapper.R
\name{buildExpSIRmatrix}
\alias{buildExpSIRmatrix}
\title{Combine all SIRratio values in a same matrix.}
\usage{
buildExpSIRmatrix(saveDirs)
}
\arguments{
\item{saveDirs}{A vector with all sample directories to be combined.}
}
\value{
A numeric matrix with SIRratio values per samples. 
Introns are in rows, samples in columns.
}
\description{
This function creates a matrix of SIRratio
with all samples in an experiment.
}
