% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_callIR.R
\name{callIRevents}
\alias{callIRevents}
\title{Call Intron Retention events from RNA long read alignments}
\usage{
callIRevents(bam, intronGR, flankWidth = 20, keepSecondaryAlignments = FALSE)
}
\arguments{
\item{bam}{Either a path to the long read alignments (bam file) from which to call IR events,
or a GappedReads object.}

\item{intronGR}{A \code{GRanges} object listing intron genomic intervals to analyse (eg: generated using ['createIntronGR'])}

\item{flankWidth}{Minimal overlap width read should have over both exon-intron junctions.
Should be greater than 0.
(Default value: 20 bp - that is 10bp on each side of the junction position)}

\item{keepSecondaryAlignments}{Whether secondary alignments should be considered (Default: FALSE).}
}
\value{
A \code{data.frame} listing aligned long read overlapping at least one intron.
}
\description{
This function calls intron retention events from long read data
by crossing reference intron intervals and long read alignments.
It will:

1. detect reads which overlap at least one intron;

2. compute the proportion of each overlapped intron that is contained in a read;

3. assess whether the read extends beyond left and right intron borders

NB: it takes into account possible gaps and splicing in read alignments.

For more details and usage see vignette('IR-detection-from-Nanopore-data')
}
\examples{

## First, create a GRanges with intron intervals from a reference transcriptome:
\dontrun{
gtf <- "Homo_sapiens.GRCh38.97.gtf"
}
## Then, call intron retention events using alignments in a bam file:
\dontrun{
bamFile <- "myLongReads.bam"
IR_events <- callIRevents(bamFile, intronGR)
}

}
