% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_LRratio.R
\name{computeIRrates}
\alias{computeIRrates}
\title{Compute observed IR rates from a Long Read experiment}
\usage{
computeIRrates(
  gtf,
  bamFile,
  saveDir,
  flankWidth = 20,
  keepSecondaryAlignments = FALSE
)
}
\arguments{
\item{gtf}{Path to a reference transcriptome is a gtf file.}

\item{bamFile}{Path to a bam file contraining long read alignments.}

\item{saveDir}{Path to the directory where results sould be saved.}

\item{flankWidth}{Width of the area on each side of the intron that a read should overlap
to be taken into account (Default: 20bp).}

\item{keepSecondaryAlignments}{Whether secondary alignments should be considered (Default: FALSE).}
}
\value{
Saves a \code{data.frame} containing observed intron retention rates
for each annotated intron in a file named "IRrates.txt" in \code{saveDir}.
}
\description{
This function computes the observed rates of intron
retention from a set of long read (genomic) alignments.
}
