% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_callIR.R
\name{computeRegionOverlap}
\alias{computeRegionOverlap}
\title{Compute the fraction of an intron contained in a long read}
\usage{
computeRegionOverlap(align.gr, intron.gr)
}
\arguments{
\item{align.gr}{A \code{GRanges} object describing a read alignment (eg: extracted from a bam file)}

\item{intron.gr}{A \code{GRanges} object describing a genomic interval to compare to \code{align.gr}}
}
\value{
The fraction of \code{intron.gr} contained in the sequence \code{align.gr}
}
\description{
This function computes the proportion
of an input genomic interval (eg: an intron or an exon) that is contained in a read.

NB: it takes into account possible gaps and splicing in read alignments.
}
