% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_SIRratio_wrapper.R
\name{computeSIRratio}
\alias{computeSIRratio}
\title{Intron retention levels estimation using short read RNA-seq data}
\usage{
computeSIRratio(
  gtf,
  keepAnnotatedRI = FALSE,
  chainPath = NULL,
  mappabilityFile = NULL,
  bamFile,
  readLength,
  libraryType,
  junctionFile,
  min_map_cross_junc = 5,
  min_novel_intron_width = 30,
  saveDir,
  verbose = TRUE
)
}
\arguments{
\item{gtf}{Reference transcriptome in a gtf/gff file. Used to define introns.}

\item{keepAnnotatedRI}{Whether intron known as retained in a transcript should be considered (Default: TRUE).}

\item{chainPath}{(Default: NULL)}

\item{mappabilityFile}{(eg: wgEncodeCrgMapabilityAlign100mer.bigWig)}

\item{bamFile}{Path to STAR alignments in a bam file.}

\item{readLength}{Length of input short reads.}

\item{libraryType}{Type of the library of RNA-seq data. Either "SE" (single-end) or "PE" (paired-end).}

\item{junctionFile}{Junction file generated by STAR from the RNA-seq sample.}

\item{min_map_cross_junc}{Minimal number of supporting reads an intron detected from splice-aware alignments should have (Default: 3).}

\item{min_novel_intron_width}{Minimal width an intron detected from splice-aware alignments should have (Default: 30).}

\item{saveDir}{Name for a Project Directory where intermediate and result files will be saved.}

\item{verbose}{(Default: TRUE)}
}
\value{
A \code{data.frame} with independent intron intervals and their IRratio2.
}
\description{
This is a wrapper function for computing SIRratio values.
It performs all steps required to estimate Intron Retention levels using RNA-seq read alignments and annotation data.
}
