% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_createIntronGR.R
\name{createIntronGR}
\alias{createIntronGR}
\title{Deduce independent intron intervals from a transcriptome annotation}
\usage{
createIntronGR(gtf)
}
\arguments{
\item{gtf}{Path to a reference transcriptome in gtff/gff format}
}
\value{
A \code{GRanges} object with independent intron intervals
}
\description{
This function extracts introns from a transcriptome annotation
}
