% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_defineIntrons.R
\name{defineIntronicIntervals}
\alias{defineIntronicIntervals}
\title{Internal function to define sample-specific intronic intervals for read counting}
\usage{
defineIntronicIntervals(
  gtf,
  saveDir,
  junctionFile,
  keepAnnotatedRI = FALSE,
  min_map_cross_junc = 5,
  min_novel_intron_width = 30,
  verbose = TRUE
)
}
\arguments{
\item{gtf}{Path towards a reference transcriptome (gtf or gff format)}

\item{saveDir}{Path to where output files will be saved}

\item{junctionFile}{Junction file output by STAR (.tab files)}

\item{keepAnnotatedRI}{Whether introns annotated as "retained" in the gtf should be considered.
(ie: transcripts annotated as "retained_intron" will be removed prior to defining intronic intervals)
(Default: TRUE)}

\item{min_map_cross_junc}{Minimal number of reads supporting a junction to consider a novel intron}

\item{min_novel_intron_width}{Minimal width for an intronic interval to be kept}

\item{verbose}{(Default: TRUE)}
}
\value{
Outputs curated and independent intronic intervals
}
\description{
This functions determines sample-specific intron borders used
for counting intronic reads.
}
