% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_SIRratio.R
\name{estimateSIRratio}
\alias{estimateSIRratio}
\title{Internal function to estimate the SIRratio}
\usage{
estimateSIRratio(saveDir, readLength)
}
\arguments{
\item{saveDir}{Path to the directory where sample results are saved.}

\item{readLength}{Total read length.}
}
\value{
Saves a \code{data.frame} containing SIRratio estimates
for each intron group in a file named "SIRratio.txt" in \code{saveDir}.
}
\description{
This function merges intronic and exonic
abundances and compute the IRratio2.
}
