% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_curateIntrons.R
\name{liftItOver}
\alias{liftItOver}
\title{Internal function for genome liftover}
\usage{
liftItOver(chainPath, oldIntervals)
}
\arguments{
\item{chainPath}{A \code{Chain} object, usually imported with ['import.chain'], or something coercible to one.}

\item{oldIntervals}{A \code{GRanges} object with genomic intervals to lift over.}
}
\value{
If out_file=NULL, a GRanges object containing intervals after lift over;
        otherwise the GRanges object will be saved in out_file.
}
\description{
This function performs genome lift over.
It is simply a wrapper for ['rtracklayer::liftover'].
}
