% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_SIRratio_wrapper.R
\name{mergebyIntron}
\alias{mergebyIntron}
\title{Merging SIRratio by reference independent intron.}
\usage{
mergebyIntron(saveDir, save = TRUE)
}
\arguments{
\item{saveDir}{Name for the Sample Directory where files output by computeSIRratio() are saved.}

\item{save}{Whether the result should be saved in a file (TRUE) or output in R as a \code{data.frame}.}
}
\value{
If save=FALSE, \code{data.frame} with independent intron intervals and their SIRratio.
If save=TRUE (default) the \code{data.frame} will be save in \code{saveDir}, in a file called "ResultsByIntron.txt".
}
\description{
This function merges SIRratio values
by reference independent intron.
}
