% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_defineIntrons.R
\name{mergeJunctionFiles}
\alias{mergeJunctionFiles}
\title{Internal function for merging introns detected by STAR}
\usage{
mergeJunctionFiles(
  junctionFiles,
  min_map_cross_junc = 5,
  min_novel_intron_width = 30,
  verbose = TRUE
)
}
\arguments{
\item{junctionFiles}{A character vector with STAR output files to be merged (.tab)}

\item{min_map_cross_junc}{Minimal number of reads supporting a junction to consider a novel intron}

\item{min_novel_intron_width}{Minimal width for an intronic interval to be kept}

\item{verbose}{(Default TRUE) Whether to print summary messages}
}
\value{
A data.frame containing intronic intervals detected by STAR
        and junction counts summarised over all input junction files
}
\description{
This junctions merges junction files
}
\seealso{
* ['defineIntronicIntervals()'] to define the intronic intervals (...)
}
