% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_prepareJunctionSites.R
\name{prepareJunctionSites}
\alias{prepareJunctionSites}
\title{Internal function for extracting junction counts from a STAR junction file}
\usage{
prepareJunctionSites(
  gtf,
  saveDir,
  junctionFile,
  min_map_cross_junc = 5,
  indIntronsFile = NULL,
  annotIntronsFile = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gtf}{Path to a reference transcriptome in a gtf/gff file.}

\item{saveDir}{Path to a Project directory.}

\item{junctionFile}{Junction file output by STAR.}

\item{min_map_cross_junc}{Minimal number of supporting reads an intron detected from splice-aware alignments should have (Default: 3).}

\item{indIntronsFile}{Path to a file with independent intron intervals as output by ['defineIntrons()'] (Default: NULL).
Need not be specified when \code{saveDir} is given and contains a file ""independent_introns.txt".}

\item{annotIntronsFile}{Path to a file with annotated intron intervals as output by ['defineIntrons()'] (Default: NULL).
Need not be specified when \code{saveDir} is given and contains a file ""all_introns.txt".}

\item{verbose}{(DEFAULT: TRUE)}
}
\value{
Creates and saves files "junction_clusters.txt" and "junction_sites.txt" in \code{saveFolder}.
}
\description{
This function extracts and merges junctions counts information (for each intron_group) from a STAR junction file.
It is intended to be used after calling ['defineIntrons()'], with the same \code{saveDir}.
}
