% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim_SIRratio_wrapper.R
\name{prepareSIRmatrix}
\alias{prepareSIRmatrix}
\title{Prepare a matrix of SIRratio values for usual primary analyses
(eg: PCA, heatmap)}
\usage{
prepareSIRmatrix(X)
}
\arguments{
\item{X}{A numeric matrix of SIRratio values per sample (columns)}
}
\value{
A numeric matrix of logit-transformed SIRratio values.
}
\description{
This function processes any SIRratio matrix
to make it for for usual primary analyses.
Namely, introns with 0 or 1 values will be removed and remaining
values will be transformed with the logit function.
}
